requirements.txt
Flask==3.0.0
openai==1.93.0
python-dotenv==1.0.0
PyPDF2==3.0.1
chromadb-1.0.15
numpy==1.26.2
pandas==2.1.4
setuptools==69.0.3
sentence-transformers==5.0.0
torch==2.7.1
transformers==4.53.1
scikit-learn